jebnf-core/src/test/java/de/weltraumschaf/jebnf/ast/nodes/ChoiceNodeTest.java
/*
 * LICENSE
 *
 * "THE BEER-WARE LICENSE" (Revision 42):
 * "Sven Strittmatter" <weltraumschaf@googlemail.com> wrote this file.
 * As long as you retain this notice you can do whatever you want with
 * this stuff. If we meet some day, and you think this stuff is worth it,
 * you can buy me a beer in return.
 *
 */

package de.weltraumschaf.jebnf.ast.nodes;

import de.weltraumschaf.jebnf.ast.nodes.ChoiceNode;
import de.weltraumschaf.jebnf.ast.Node;
import de.weltraumschaf.jebnf.ast.NodeType;
import static org.junit.Assert.assertEquals;
import org.junit.Test;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

/**
 *
 * @author Sven Strittmatter <weltraumschaf@googlemail.com>
 */
public class ChoiceNodeTest {

    @Test public void testToString() {
        final ChoiceNode choice = ChoiceNode.newInstance();
        assertEquals("<CHOICE>", choice.toString());
        final Node node1 = mock(Node.class);
        when(node1.toString()).thenReturn("<foo>");
        choice.addChild(node1);
        final Node node2 = mock(Node.class);
        when(node2.toString()).thenReturn("<bar>");
        choice.addChild(node2);
        assertEquals("<CHOICE>\n<foo>\n<bar>", choice.toString());
    }

    @Test public void getType() {
        assertEquals(NodeType.CHOICE, ChoiceNode.newInstance().getType());
    }

}
